requirements.txt
alembic==1.9.1
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.6.3
Babel==2.11.0
cachetools==4.2.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
packaging==23.0
Pillow==9.4.0
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
speaklater==1.3
SQLAlchemy==1.4.46
telegram-send==0.34
tornado==6.1
tzdata==2022.7
tzlocal==4.2
Werkzeug==2.2.2
WTForms==3.0.1
